Author: Perry Lorier <isomer@undernet.org> Log message:
Lots of fixes to the convert-conf program.  It no longer generates Server
blocks (which have been removed from the latest copy of the config), it's
smarter about how it generates it's I:line configs, the line parser now
is much closer to ircu's parser avoiding a lot of bugs.  And a multitude
of other minor problems fixed up.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@1318 c9e4aea6-c8fd-4c43-8297-357d70d61c8c