Author: Kev <klmitch@mit.edu> Log message:
Get rid of MAGIC_REMOTE_JOIN_TS--hasn't been needed for a long time, we've
just been lazy.  Also, fix two bugs having to do with remote users joining
channels: 1) If the timestamps were the same, everybody got deopped; 2) the
deops weren't actually happening, just shown.  Now, the deops actually
happen, and the deop loop is only entered if the remote join TS is older
(less than) the channel timestamp or if the channel is a zannel.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_12_branch@1580 c9e4aea6-c8fd-4c43-8297-357d70d61c8c