Author: Kev <klmitch@mit.edu> Log message:
Initial work to update the command syntax and capabilities of G-lines.

*NOTE*NOTE*NOTE*NOTE*NOTE*NOTE*NOTE*NOTE*NOTE*NOTE*NOTE*NOTE*NOTE*NOTE*NOTE*

While ircu will compile and run, GLINES WILL NOT FUNCTION AS EXPECTED!
This commit inserts debugging code that will wallops what actions would be
taken to implement a particular G-line received from a remote server, but
will not actually take those actions as of yet.

DO NOT RUN THIS CODE ON A PRODUCTION NETWORK!  Thank you :)


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_12_branch@1774 c9e4aea6-c8fd-4c43-8297-357d70d61c8c