Author: Kev <klmitch@mit.edu> Log message:
Finish changes to G-line.  These changes enable "/stats g" and "/gline" to
show G-line lastmod, lifetime, and local activation status, as well as
allowing G-lines to be modified in virtually any way imaginable.  Note that
overlapping G-lines are now to be kept, as a potentially masking G-line may
actually be disabled and so not actually be masking a new G-line...


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_12_branch@1780 c9e4aea6-c8fd-4c43-8297-357d70d61c8c