Author: Kev <klmitch@mit.edu> Log message:
This very simple change adds a +R channel mode that can only be set or
reset by remote users (or servers) or with /opmode.  This can be used by X
to indicate that a channel is registered, a sentinal that C can use to
decide when it is legal to reop a channel.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_12_branch@1860 c9e4aea6-c8fd-4c43-8297-357d70d61c8c