Author: Kev <klmitch@mit.edu> Log message:
Creation of a new generic registration subsystem.  The purpose is basically
to allow dynamic addition of entries for formerly table-driven stuff, like
/stats; this will enable us to respect abstraction boundaries much better.
For instance, the /stats subcommands for iauth could be declared in
s_auth.c itself, instead of having to make the called functions visible to
s_stats.c.

Note that this is only the implementation, and that the only tests done so
far are to check that it compiles.  It is not added to the sources list,
and nothing in the core code has yet been modified to make use of the new
facility.  This is part of the modularization effort for ircu2.10.13.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@1873 c9e4aea6-c8fd-4c43-8297-357d70d61c8c