Author: Kev <klmitch@mit.edu> Log message:
Creation of a new generic ancillary data subsystem.  The purpose is to
allow other modules to add their own data to objects such as clients, while
respecting abstraction boundaries.  The design is inspired by the POSIX
threads pthread_{get,set}specific() interface.

Note that this is only the implementation, and that the only tests done so
far are to check that it compiles.  It is not added to the sources list,
and nothing in the core code has yet been modified to make use of the new
facility.  This is part of the modularization effort for ircu2.10.13.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@1874 c9e4aea6-c8fd-4c43-8297-357d70d61c8c