Author: Kev <klmitch@mit.edu> Log message:
Creation of a new generic "watches" subsystem.  The purpose is to allow a
module to notify other modules of certain "events" on its data structures;
examples might include notification of the creation of a user or the
destruction of a channel, or even the modification of some field in the
data structure.  This adds the possibility of dynamic enhancement involving
watching for changes in a data structure, and improves abstraction
boundaries.

Note that this is only the implementation, and that the only tests done so
far are to check that it compiles.  It is not added to the sources list,
and nothing in the core code has yet been modified to make use of the new
facility.  This is part of the modularization effort for ircu2.10.13.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@1875 c9e4aea6-c8fd-4c43-8297-357d70d61c8c