Author: Kev <klmitch@mit.edu> Description:
Begin sketching out the structures and fields necessary for the new,
unified mode processing system.  Each entity (channel, user, or server)
will have a mode_set_t (modes with arguments will set those arguments
in fields directly adjunct to the entity structure, rather than collected
into the mode_set_t).  Each mode is described by a mode_desc_t, and the
list of modes for a given entity are described by a mode_list_t.
Processing will consist of parsing the mode string into a mode_delta_t,
then applying it to the entity.  (Bouncing will be simple; set the
MDELTA_REVERSE flag on the mode_delta_t before flushing it.)

Modes and mode lists will be linked in to the registration system,
making it easy to dynamically define new entities that can take modes
(e.g., services), as well as to dynamically define new modes for any
entity.  Manipulating a mode will require a pointer to the mode_desc_t
for the mode, instead of the integer or flag value of the mode, so that
the mode system can dynamically assign that integer.

Couple of notes on terminology: "switch" specifies the character for
the mode, i.e., the no-topic switch is 't'.  The "mode" is the integer
specifying the mode (the mode_t type is used for this).  Modes also have
full names, allowing a future extension where you can use long names for
modes instead of just single characters.



git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/mode@1880 c9e4aea6-c8fd-4c43-8297-357d70d61c8c