Author: Kev <klmitch@mit.edu> Description:
Begin roughing in the interface for manipulating deltas.  This includes
adding the concept of "destinations," which control where the mode
gets delivered.  (This plays an equivalent role to the MODEBUF_DEST_*
constants used with the existing modebuf system.)

I also add a few points; for instance, keys are only visible to channel
operators, but when an explicit MODE is used to set a key, we allow that
key to be seen by all channel users, since we know they're all
legitimate.  Thus, I added the MDFLAG_SHOWARG flag for mode descriptors.
Further, +A and +U allow their passwords to be changed, even though you
can't remove them by simply allowing, e.g., -A+A <oldpass> <newpass>;
the MDFLAG_ONECHANGE flag was added to complement the MDFLAG_ONESHOT
flag and allow this behavior.

Finally, we were starting to run out of bits in the flags element of
mode_desc_t, so I packed the indicators a little more tightly--and to
keep the clarity, I added some macros to apply the appropriate bit shift
to them.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/mode@1890 c9e4aea6-c8fd-4c43-8297-357d70d61c8c