Fix usermodes harder when they are set early during registration.
IAuth and /user can both set usermodes before registration is
complete, and we do not want to count unregistered clients as
invisible or as opers, so register_user() _must_ update UserStats when
it registers a client.  Default usermodes must not be applied between
where the client is marked as registered and that point, because that
would update UserStats too, so move the setting of default usermodes
before the SetUser() call that marks the client as a registered user.

Fixes SourceForge bugs #2824908 and #2829936.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_12_branch@1919 c9e4aea6-c8fd-4c43-8297-357d70d61c8c