s_auth: Check G-lines before IAuth.
Add a new function, iauth_notify(), that sends the appropriate message
to the IAuth service when some information about a client is updated.

check_auth_finished() now calls this function, and as a result needs to
know what information was just updated.  However, it also calls
register_user() before sending the last required information to IAuth,
so that a G-line keeps the IAuth service from performing full checks.

The syntax for the U message to IAuth is updated to remove fields that
ircu interprets as mode requests.