s_auth: Fix prepending ~ to long idents.
This was touched in commits 5dd977c1dc9f33ddfbca5eb11fb8f3b95c7e5359 and
c2ad630afaab5353b2656f6cd5d591f8a2488692, but still had a bug.  We can
rewrite the loop to make it clearer, simpler, and harder to mess up.

Thanks to Hidden for reporting the bug and Scot for finding the cause.