s_auth: Replace FLAG_DOID with DoIdentLookups.
Ever since commit bdd001c7946:
  s_auth: Let IAuth assign class without connection-limit or password checks.
we have called preregister_user() (and indirectly attach_iline()) after
all the other authorization checks are done.  This means that FLAG_DOID
was being set after it was being checked.

There was a Gordian knot:
 - We want IAuthd to be able to set a connection class.
 - We want to check G-lines before sending enough information to IAuthd
   to trigger an XQUERY lookup (meaning network traffic).
 - We use clients' usernames to check G-lines.
 - We add ~ to the client's username if their connection class needs a
   username, but the identd lookup failed for that cilent.

I think the simplest, least-surprising way to break that loop is to add a
~ prefix if *any* class needs a username.  Most servers will continue to
behave the same as with u2.10.12.18, and this makes the server's behavior
more consistent across clients who fall into different connection classes.