Support primitive parser generators like byacc
yacc expects zero or negative from yylex() on EOF, but not all
implementations define YYEOF in y.tab.h.  Define it if needed.

And most variants of yacc seem to use 256 as an error token, but that is
not specified and not always exported.  So define a new token (TOKERR)
to indicate a tokenization error.

Fixes #15