Use 336/337 instead of 346/347 to reply to 'INVITE' commands
RFC2812 defines `RPL_INVITELIST` (346) and `RPL_ENDOFINVITELIST` (347) numerics,
but they are only to be used for replies to 'MODE +I', which stands for
'invite exemption' (though it is, surprisingly, not defined in RFC2812
itself). An they must have two params: the channel and mask
<https://datatracker.ietf.org/doc/html/rfc2812#page-46>.

Instead, the consensus is to use 336 and 337 to reply to parameter-less
INVITE messages; and call these `RPL_INVITELIST`/`RPL_ENDOFINVITELIST`,
while the RFC2812 numerics should be renamed to something like
`RPL_INVEXLIST`/`RPL_ENDOFINVEXLIST` (which ircu2 does not use, so it
does not matter much). See for example:

* https://github.com/inspircd/inspircd/commit/df17d47b6a17ee6214f7f501e3b9d73cb8acd36e
* https://github.com/unrealircd/unrealircd/commit/a11e6df64b9ab8fe27ecbc1300893bf8796dcebc