Fix bug that rejects valid usernames with two digit groups
The code is supposed to reject usernames that contain two digit groups, unless
one of the groups is at the start or end of the username. However, at the end
of the loop, 'ch' is the null byte, so usernames that ended with a second
digit group were being rejected.

Fix this by checking 'last' instead of 'ch'.